Changed default
In line 307, s is set to the integer version of the number in the string, if the string is not a number it should default to 0 (standard webcam for cv2.VideoCapture() ), however it is set to s which in this case would be a string, which would give an error in the next line. This is just stupid so changed it to i (which is always 0).

The fact that i even exists due to an enumerate for loop which only ever loops once is stupid but ill not make any drastic changes to the code, just ranting. 

# :)